locale.c: Enable shortened critical section times
Since 5335601b6b6a88f5244d09fd2c1e531a1918ec12, there has been the
capability of having nested locks, where the inner one is necessary in
all conditions, and the outer one is desirable in just some conditions.

But this hasn't been enabled because in some Configurations it could
cause a deadlock, with the thread locking the ENV mutex, and then in a
nested call trying to lock it again.  The previous commit made this mutex reentrant, so the deadlock is gone.