Move locale mutext init/destruct
I can't find a place in this file for this to go that isn't somewhat
awkward.  Because the environment and locale mutexes are intertwined,
they naturally fit together.  One would think the basic initialization
and termination of the locale mutexes should go ahead of their use.  But
because of all the complications, that interrupts the flow of the logic
between the environment and locale mutexes, so they are moved to later
by this commit.